+++ lib/processor/metric.go
@@ -145,10 +145,15 @@ type Metric struct {
 
 	interpolateValue bool
 
-	labels   labels
-	mCounter metrics.StatCounterVec
-	mGauge   metrics.StatGaugeVec
-	mTimer   metrics.StatTimerVec
+	labels labels
+
+	mCounter metrics.StatCounter
+	mGauge   metrics.StatGauge
+	mTimer   metrics.StatTimer
+
+	mCounterVec metrics.StatCounterVec
+	mGaugeVec   metrics.StatGaugeVec
+	mTimerVec   metrics.StatTimerVec
 
 	handler func(string, types.Message) error
 }
@@ -215,19 +220,39 @@ func NewMetric(
 
 	switch strings.ToLower(conf.Metric.Type) {
 	case "counter":
-		m.mCounter = stats.GetCounterVec(conf.Metric.Path, m.labels.names())
+		if len(m.labels) > 0 {
+			m.mCounterVec = stats.GetCounterVec(conf.Metric.Path, m.labels.names())
+		} else {
+			m.mCounter = stats.GetCounter(conf.Metric.Path)
+		}
 		m.handler = m.handleCounter
 	case "counter_parts":
-		m.mCounter = stats.GetCounterVec(conf.Metric.Path, m.labels.names())
+		if len(m.labels) > 0 {
+			m.mCounterVec = stats.GetCounterVec(conf.Metric.Path, m.labels.names())
+		} else {
+			m.mCounter = stats.GetCounter(conf.Metric.Path)
+		}
 		m.handler = m.handleCounterParts
 	case "counter_by":
-		m.mCounter = stats.GetCounterVec(conf.Metric.Path, m.labels.names())
+		if len(m.labels) > 0 {
+			m.mCounterVec = stats.GetCounterVec(conf.Metric.Path, m.labels.names())
+		} else {
+			m.mCounter = stats.GetCounter(conf.Metric.Path)
+		}
 		m.handler = m.handleCounterBy
 	case "gauge":
-		m.mGauge = stats.GetGaugeVec(conf.Metric.Path, m.labels.names())
+		if len(m.labels) > 0 {
+			m.mGaugeVec = stats.GetGaugeVec(conf.Metric.Path, m.labels.names())
+		} else {
+			m.mGauge = stats.GetGauge(conf.Metric.Path)
+		}
 		m.handler = m.handleGauge
 	case "timing":
-		m.mTimer = stats.GetTimerVec(conf.Metric.Path, m.labels.names())
+		if len(m.labels) > 0 {
+			m.mTimerVec = stats.GetTimerVec(conf.Metric.Path, m.labels.names())
+		} else {
+			m.mTimer = stats.GetTimer(conf.Metric.Path)
+		}
 		m.handler = m.handleTimer
 	default:
 		return nil, fmt.Errorf("metric type unrecognised: %v", conf.Metric.Type)
@@ -237,7 +262,11 @@ func NewMetric(
 }
 
 func (m *Metric) handleCounter(val string, msg types.Message) error {
-	m.mCounter.With(m.labels.values(msg)...).Incr(1)
+	if len(m.labels) > 0 {
+		m.mCounterVec.With(m.labels.values(msg)...).Incr(1)
+	} else {
+		m.mCounter.Incr(1)
+	}
 	return nil
 }
 
@@ -245,7 +274,11 @@ func (m *Metric) handleCounterParts(val string, msg types.Message) error {
 	if msg.Len() == 0 {
 		return nil
 	}
-	m.mCounter.With(m.labels.values(msg)...).Incr(int64(msg.Len()))
+	if len(m.labels) > 0 {
+		m.mCounterVec.With(m.labels.values(msg)...).Incr(int64(msg.Len()))
+	} else {
+		m.mCounter.Incr(int64(msg.Len()))
+	}
 	return nil
 }
 
@@ -257,7 +290,11 @@ func (m *Metric) handleCounterBy(val string, msg types.Message) error {
 	if i < 0 {
 		return errors.New("value is negative")
 	}
-	m.mCounter.With(m.labels.values(msg)...).Incr(i)
+	if len(m.labels) > 0 {
+		m.mCounterVec.With(m.labels.values(msg)...).Incr(i)
+	} else {
+		m.mCounter.Incr(i)
+	}
 	return nil
 }
 
@@ -269,7 +306,11 @@ func (m *Metric) handleGauge(val string, msg types.Message) error {
 	if i < 0 {
 		return errors.New("value is negative")
 	}
-	m.mGauge.With(m.labels.values(msg)...).Set(i)
+	if len(m.labels) > 0 {
+		m.mGaugeVec.With(m.labels.values(msg)...).Set(i)
+	} else {
+		m.mGauge.Set(i)
+	}
 	return nil
 }
 
@@ -281,7 +322,11 @@ func (m *Metric) handleTimer(val string, msg types.Message) error {
 	if i < 0 {
 		return errors.New("value is negative")
 	}
-	m.mTimer.With(m.labels.values(msg)...).Timing(i)
+	if len(m.labels) > 0 {
+		m.mTimerVec.With(m.labels.values(msg)...).Timing(i)
+	} else {
+		m.mTimer.Timing(i)
+	}
 	return nil
 }